[OGUI-755] Split test file in multiple ones
