+++ InfoLogger/test/public/mocha-live-mode.js
@@ -1,0 +1,160 @@
+/**
+ * @license
+ * Copyright 2019-2020 CERN and copyright holders of ALICE O2.
+ * See http://alice-o2.web.cern.ch/copyright for details of the copyright holders.
+ * All rights not expressly granted are reserved.
+ *
+ * This software is distributed under the terms of the GNU General Public
+ * License v3 (GPL Version 3), copied verbatim in the file "COPYING".
+ *
+ * In applying this license CERN does not waive the privileges and immunities
+ * granted to it by virtue of its status as an Intergovernmental Organization
+ * or submit itself to any jurisdiction.
+*/
+
+/* eslint-disable max-len */
+
+const assert = require('assert');
+const test = require('../mocha-index');
+
+describe('Live Mode test-suite', async () => {
+  let baseUrl;
+  let page;
+
+  before(async () => {
+    baseUrl = test.helpers.baseUrl;
+    page = test.page;
+    // TODO reload page
+  });
+  describe('Live mode', async () => {
+    it('can be activated because it is configured and simulator is started', async () => {
+      const activeMode = await page.evaluate(() => {
+        window.model.log.liveStart();
+        return window.model.log.activeMode;
+      });
+  
+      assert.deepStrictEqual(activeMode, 'Running');
+    });
+  
+    it('cannot be activated twice', async () => {
+      const thrown = await page.evaluate(() => {
+        try {
+          window.model.log.liveStart();
+          return false;
+        } catch (e) {
+          return true;
+        }
+      });
+  
+      assert.deepStrictEqual(thrown, true);
+    });
+  
+    it('should have filled some logs via WS with the level "debug"', async () => {
+      // check level is still 21 after LogFilter tests
+      const criterias = await page.evaluate(() => {
+        window.model.log.filter.resetCriterias();
+        window.model.log.filter.setCriteria('level', 'max', 21);
+        return window.model.log.filter.criterias;
+      });
+  
+      assert.deepStrictEqual(criterias.level.max, 21);
+      assert.deepStrictEqual(criterias.level.$max, 21);
+  
+      // Wait for logs and count them (2-3 maybe, it's random)
+      await page.waitForTimeout(1500); // simulator is set to ~100ms per log
+      const list = await page.evaluate(() => {
+        return window.model.log.list;
+      });
+      assert.deepStrictEqual(!!list.length, true);
+    });
+  
+    it('should filter messages based on `hostname` matching `aldaqecs01-v1` from live -> paused -> live', async () => {
+      await page.evaluate(() => window.model.log.liveStop('Paused'));
+      await page.evaluate(() => {
+        window.model.log.filter.resetCriterias();
+        window.model.log.filter.setCriteria('hostname', 'match', 'aldaqecs01-v1');
+      });
+      await page.evaluate(() => window.model.log.liveStart());
+      await page.waitForTimeout(7000);
+      const list = await page.evaluate(() => window.model.log.list);
+      await page.waitForTimeout(1000);
+      const isHostNameMatching = list.map((element) => element.hostname).every((hostname) => hostname === 'aldaqecs01-v1');
+      assert.ok(list.length > 0);
+      assert.ok(isHostNameMatching);
+    });
+  
+    it('should filter messages based on `hostname` excluding `aldaqdip01` from live -> query -> live', async () => {
+      await page.evaluate(() => window.model.log.liveStop('Query'));
+      await page.evaluate(() => {
+        window.model.log.filter.resetCriterias();
+        window.model.log.filter.setCriteria('hostname', 'exclude', 'aldaqdip01');
+      });
+      await page.evaluate(() => window.model.log.liveStart());
+      await page.waitForTimeout(3000);
+      const list = await page.evaluate(() => window.model.log.list);
+      const isHostNameMatching = list.map((element) => element.hostname).every((hostname) => hostname !== 'aldaqdip01');
+  
+      assert.ok(list.length > 0);
+      assert.ok(isHostNameMatching);
+    });
+  
+    it('should filter messages based on SQL Wildcards `hostname` excluding `%ldaqdip%` and username matching `a_iceda_` without changing state of live mode', async () => {
+      await page.evaluate(() => window.model.log.filter.resetCriterias());
+      await page.evaluate(() => {
+        window.model.log.setCriteria('hostname', 'exclude', '%ldaqdip%');
+        window.model.log.setCriteria('username', 'match', 'a_iceda_');
+        window.model.log.empty();
+      });
+      await page.waitForTimeout(3000);
+      const list = await page.evaluate(() => window.model.log.list);
+      const isHostNameMatching = list.map((element) => element.hostname).every((hostname) => !new RegExp('.*ldaqdip.*').test(hostname));
+      const isUserNameMatching = list.map((element) => element.username).every((username) => new RegExp('a.iceda.').test(username));
+  
+      assert.ok(list.length > 0);
+      assert.ok(isHostNameMatching);
+      assert.ok(isUserNameMatching);
+    });
+  
+    it('successfully show indicator when user double pressed the log row', async () => {
+      const tableRow = await page.$('body > div:nth-child(2) > div:nth-child(2) > main > div > div > div > table > tbody > tr');
+      await tableRow.click({clickCount: 2});
+      await page.waitForTimeout(200);
+      const indicatorOpen = await page.evaluate(() => window.model.inspectorEnabled);
+      assert.ok(indicatorOpen);
+    });
+  
+    it('should go to mode live in paused state', async () => {
+      const activeMode = await page.evaluate(() => {
+        window.model.log.liveStop('Paused');
+        return window.model.log.activeMode;
+      });
+  
+      assert.deepStrictEqual(activeMode, 'Paused');
+    });
+  
+    it('should go to mode query', async () => {
+      const activeMode = await page.evaluate(() => {
+        window.model.log.liveStart();
+        window.model.log.liveStop('Query');
+        return window.model.log.activeMode;
+      });
+  
+      assert.deepStrictEqual(activeMode, 'Query');
+    });
+  
+    it('should go to mode query if mode not specified', async () => {
+      const activeMode = await page.evaluate(() => {
+        window.model.log.liveStart();
+        window.model.log.liveStop();
+        return window.model.log.activeMode;
+      });
+  
+      assert.deepStrictEqual(activeMode, 'Query');
+    });
+  });
+});
+
+
+
+
+
